# Optimize scans by using noinit for the result array