… test/scan/testScanPOD.good
… test/scan/testScanPOD.numlocales