@bradcray 2021-01-13
This looks good to me too (except that now I want noinit of non-POD arrays... :D  ).